refactor: 구글과 애플 OAuth 관련 코드 분리
